Add support for random access to the blobstore API.
The BlobAccess interface only provides support for streaming files as a
whole. This is desired, because it also allows us to do checksum
validation using the MerkleBlobAccess.

If we want to provide support for running build actions inside of a FUSE
file system, we'll need to add logic for obtaining file contents
partially and at arbitrary offsets. Add a RandomAccessBlobAccess that is
a superset of a normal BlobAccess, having a GetAndReadAt function,
inspired by io.ReaderAt.